refactor cost table
