+++ frontend/src/types/costTable.ts
@@ -1,0 +1,15 @@
+export type DataRow = {
+  id: number;
+
+  name: string;
+  processor: string;
+  cpu: number;
+  memory: string;
+  leaseLength: string;
+  region: string;
+  engineType: string;
+  commitment: { usd: number };
+  hourly: { usd: number };
+
+  childCnt: number;
+};
