+++ frontend/src/types/index.ts
@@ -1,5 +1,6 @@
 export * from "./contributor";
 export * from "./dbInstance";
+export * from "./costTable";
 export * from "./region";
 export * from "./term";
 export * from "./common";
