+++ frontend/src/views/AppView.vue
@@ -15,6 +15,7 @@
     >
     <span>💸</span>
   </h1>
+
   <!-- menu -->
   <div class="mx-5 mt-4">
     <div class="mb-4 justify-between flex">
@@ -44,20 +45,22 @@
   <!-- dashboard -->
   <div class="mx-5 mt-5">
     <cost-table
-      :db-instance-list="dbInstanceStore.dbInstanceList"
-      :config="searchConfigStore.searchConfig"
+      :data-row="state.dataRow"
+      :isLoading="state.isLoading"
+      :show-engine-type="showEngineType"
+      :show-lease-length="showLeaseLength"
     />
   </div>
 </template>
 
 <script setup lang="ts">
-import { watch } from "vue";
+import { watch, reactive, ref, computed } from "vue";
 
 import CostTable from "../components/CostTable.vue";
 import CostTableMenu from "../components/CostTableMenu.vue";
 import { NButton, useNotification } from "naive-ui";
 
-import { ChargeType, DBInstance, EngineType } from "../types";
+import { ChargeType, DBInstance, EngineType, DataRow } from "../types";
 import { useDBInstanceStore } from "../stores/dbInstance";
 import { useSearchConfigStore } from "../stores/searchConfig";
 import { useRouter } from "vue-router";
@@ -70,6 +73,16 @@ const dbInstanceStore = useDBInstanceStore();
 dbInstanceStore.dbInstanceList = aws as unknown as DBInstance[];
 
 const searchConfigStore = useSearchConfigStore();
+
+interface LocalState {
+  dataRow: DataRow[];
+  isLoading: boolean;
+}
+
+const state = reactive<LocalState>({
+  dataRow: [],
+  isLoading: false,
+});
 
 const handleUpdateRegion = (val: string[]) => {
   searchConfigStore.searchConfig.region = val;
@@ -139,6 +152,140 @@ const copyURL = () => {
         duration: 2000,
       });
     });
+};
+
+const config = ref(searchConfigStore.searchConfig);
+watch(
+  () => config, // ref to searchConfigStore.searchConfig
+  () => {
+    state.isLoading = true;
+    refreshDataTable();
+    state.isLoading = false;
+  },
+  {
+    deep: true,
+  }
+);
+
+const showEngineType = computed(
+  () => config.value.engineType && config.value.engineType.length > 1
+);
+const showLeaseLength = computed(() => {
+  const chargeTypeSet = new Set(config.value.chargeType);
+  return chargeTypeSet.size > 1 || chargeTypeSet.has("Reserved");
+});
+const refreshDataTable = () => {
+  state.dataRow = [];
+  let rowCnt = 0;
+  const config = searchConfigStore.searchConfig;
+  const dbInstanceList = dbInstanceStore.dbInstanceList;
+
+  if (!config.region && !config.engineType && !config.chargeType) {
+    return;
+  }
+
+  const selectedRegionSet = new Set(config.region);
+  const engineSet = new Set(config.engineType);
+  const chargeTypeSet = new Set(config.chargeType);
+
+  dbInstanceList.forEach((dbInstance) => {
+    if (
+      (config.minRAM && Number(dbInstance.memory) < config.minRAM) ||
+      (config.minCPU && Number(dbInstance.cpu) < config.minCPU)
+    ) {
+      return;
+    }
+
+    const selectedRegionList = dbInstance.regionList.filter((region) => {
+      if (selectedRegionSet.has(region.name)) {
+        return true;
+      }
+      return false;
+    });
+    // no region found
+    if (selectedRegionList.length === 0) {
+      return;
+    }
+
+    const dataRowList: DataRow[] = [];
+    const dataRowMap: Map<string, DataRow[]> = new Map();
+
+    // filter terms provided in each region
+    selectedRegionList.forEach((region) => {
+      const selectedTermList = region.termList.filter((term) => {
+        if (
+          chargeTypeSet.has(term.type) &&
+          engineSet.has(term.databaseEngine)
+        ) {
+          return true;
+        }
+        return false;
+      });
+
+      selectedTermList.forEach((term) => {
+        const key = `${dbInstance.name}-${region.name}`;
+        const newRow: DataRow = {
+          id: -1,
+          childCnt: 1,
+          name: dbInstance.name,
+          processor: dbInstance.processor,
+          cpu: dbInstance.cpu,
+          memory: dbInstance.memory,
+          engineType: term.databaseEngine,
+          commitment: { usd: term.commitmentUSD },
+          hourly: { usd: term.hourlyUSD },
+          leaseLength: term.payload?.leaseContractLength ?? "N/A",
+          region: region.name,
+        };
+
+        if (dataRowMap.has(key)) {
+          const existedDataRowList = dataRowMap.get(key) as DataRow[];
+          newRow.id = existedDataRowList[0].id;
+          existedDataRowList.push(newRow);
+        } else {
+          rowCnt++;
+          newRow.id = rowCnt;
+          dataRowMap.set(key, [newRow]);
+        }
+      });
+    });
+
+    dataRowMap.forEach((val) => {
+      val.sort((a, b) => {
+        // sort the row according to the following criteria
+        // 1. charge type
+        // 2. ascend by lease length
+        if (a.leaseLength == "N/A") {
+          return -1;
+        } else if (b.leaseLength == "N/A") {
+          return 1;
+        }
+        return Number(a.leaseLength[0]) - Number(b.leaseLength[0]);
+      });
+      val[0].childCnt = val.length;
+      dataRowList.push(...val);
+    });
+
+    // filter by keyword, we only enable this when the keyword is set by user
+    const keyword = config.keyword;
+    if (keyword) {
+      const filteredDataRowList: DataRow[] = dataRowList.filter((row) => {
+        if (
+          row.name.includes(keyword) ||
+          row.memory.includes(keyword) ||
+          row.processor.includes(keyword) ||
+          row.region.includes(keyword)
+        ) {
+          return true;
+        }
+        return false;
+      });
+      state.dataRow.push(...filteredDataRowList);
+      return;
+    }
+
+    state.dataRow.push(...dataRowList);
+  });
 };
 
 const clearAll = () => {
